ファイルの名前を変更してサンプリング周波数の設定ボタンを付ける
